+++ CHANGELOG.md
@@ -17,6 +17,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Add support to render template using `host` or `$(uname -n)` as the value of `overlay show --render`. #623
 - Added command line parameters for credentials of a container registry
 - Add flag `--build` to `wwctl container copy`. #1378
+- container images can be imported for different platforms using the --platform flag
 
 ### Changed